# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gerenciamento de Fórmulas de Manipulação",
        "description": "A API foi desenvolvida para uma farmácia de manipulação com o objetivo de gerenciar clientes, fórmulas de manipulação e os ativos utilizados nessas fórmulas.",
        "contact": {
            "email": "francorodrigognr@yahoo.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/formulas": {
            "get": {
                "tags": [
                    "Fórmulas"
                ],
                "summary": "Obter lista de fórmulas",
                "operationId": "763323a10846eacb4eb797deee158eda",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Formula"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Fórmulas"
                ],
                "summary": "Criar uma nova fórmula",
                "operationId": "190955cbd5f4e6310b242c776cd8d493",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "descricao",
                                    "cliente_id"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Fórmula A"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Descrição da Fórmula A"
                                    },
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ativos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fórmula criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Formula"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/formulas/{id}": {
            "get": {
                "tags": [
                    "Fórmulas"
                ],
                "summary": "Exibir detalhes de uma fórmula",
                "operationId": "0ad0a929747cf75baeac4c82464bdc99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da fórmula",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Formula"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Fórmulas"
                ],
                "summary": "Atualizar uma fórmula existente",
                "operationId": "6c6d23adbef68b7dadc56f4d3333a237",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da fórmula",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Fórmula Atualizada"
                                    },
                                    "descricao": {
                                        "type": "string",
                                        "example": "Descrição atualizada da fórmula"
                                    },
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ativos": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fórmula atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Formula"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requisição inválida"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Fórmulas"
                ],
                "summary": "Excluir uma fórmula existente",
                "operationId": "56f29936439e62ac81b67cb541ff3cec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da fórmula",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Fórmula excluída com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Formula": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Formula Example"
                    },
                    "descricao": {
                        "type": "string",
                        "example": "Description of the formula"
                    },
                    "cliente_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-14T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-08-14T10:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}